Refactored AbstractFigure and AbstractRectangularFigure
